% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_phone.R
\name{phone_number_code_from_string}
\alias{phone_number_code_from_string}
\title{Extract UK phone numbers from a string}
\usage{
phone_number_code_from_string(string)
}
\arguments{
\item{string}{A character string which may contain a phone number.}
}
\value{
A tibble with 4 columns containing the original string, a logical
indicated whether a phone number was detected, a redacted string and the redacted
terms.
}
\description{
Extract UK phone numbers from a string
}
\examples{

phone_number_code_from_string("07123123123")
phone_number_code_from_string("+447123123123")


}
