% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_pcd.R
\name{post_code_from_string}
\alias{post_code_from_string}
\title{Extract UK post codes from a string}
\usage{
post_code_from_string(string)
}
\arguments{
\item{string}{A character string which may contain a post code.}
}
\value{
A tibble with 4 columns containing the original string, a logical
indicated whether a postcode was detected, a redacted string and the redacted
terms.
}
\description{
Extract UK post codes from a string
}
\examples{

post_code_from_string("The postcode for Big Ben is SW1A 0AA.")

}
