% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_setup.r
\name{python_setup}
\alias{python_setup}
\title{python_setup}
\usage{
python_setup()
}
\description{
This function checks installs the python library spacy and the
English language model on which named entity extraction relies. You must have
python installed already for this to work.
}
