% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_all.R
\name{redact_all}
\alias{redact_all}
\title{redact_all}
\usage{
redact_all(string)
}
\arguments{
\item{string}{A character string which may contain personal data.}
}
\value{
A string with potential personal data redacted.
}
\description{
This function takes a string and redacts all the phone numbers,
email addresses, post codes, and other named entities that it can in an effort
to exclude accidentally collected personal data from subsequent analysis.
}
