% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_named_entity.R
\name{redact_named_entities}
\alias{redact_named_entities}
\title{redact_named_entities}
\usage{
redact_named_entities(string)
}
\arguments{
\item{string}{A character string which may contain a named entity.}
}
\value{
A tibble with 4 columns containing the original string, a logical
indicated whether a named entity was detected, a redacted string and the redacted
terms.
}
\description{
Relies on python_setup having run successfully.
Redact named entities e.g. people, addresses, from a string.
}
