% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redact_pcd.R
\name{redact_post_code}
\alias{redact_post_code}
\title{redact_post_code}
\usage{
redact_post_code(string)
}
\arguments{
\item{string}{A character string which may contain a post code.}
}
\value{
A tibble with 4 columns containing the original string, a logical
indicated whether a postcode was detected, a redacted string and the redacted
terms.
}
\description{
Extract UK post codes from a string.
}
\examples{

redact_post_code("The postcode for Big Ben is SW1A 0AA.")

}
